added user registration controller

+++ src/utils/cloudinary.js
@@ -1,7 +1,8 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
+import dotenv from "dotenv";
+dotenv.config(); // ensure env variables are loaded
 
-// Configuration
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
@@ -11,16 +12,21 @@ cloudinary.config({
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null;
+        //upload the file on cloudinary
         const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: auto,
+            resource_type: "auto",
         });
-        console.log(
-            "File uploaded successfully on cloudinary at -->  ",
-            response?.url
-        );
+        // file has been uploaded successfull
+        console.log("file is uploaded on cloudinary ", response.url);
+        fs.unlinkSync(localFilePath);
         return response;
     } catch (error) {
-        fs.unlinkSync(localFilePath);
+        console.error("Cloudinary upload failed:", error.message);
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        } else {
+            console.warn("File not found while cleaning up:", localFilePath);
+        }
         return null;
     }
 };
